requirements.txt
future==0.18.2
neuralpy-torch==0.1.1
numpy==1.19.0
pandas==1.0.5
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
torch==1.5.1+cpu
torchvision==0.6.1+cpu
